+++ BakeryTracker/Models/Order.cs
@@ -7,11 +7,11 @@ namespace BakeryTracker.Models
     public static List<Order> _instances = new List<Order> { };
     public string Title { get; set; }
     public string Description { get; set; }
-    public int Price { get; set; }
+    public double Price { get; set; }
     public string Date { get; set; }
     public string Frequency { get; set; }
     public int Id { get; set; }
-    public Order(string title, string description, int price, string date, string frequency)
+    public Order(string title, string description, double price, string date, string frequency)
     {
       Title = title;
       Description = description;
